LayoutTests/fast/forms/extrazoom/time-picker-value-change-expected.txt

PASS time.value is '09:41'
PASS successfullyParsed is true

TEST COMPLETE
